\name{RRedsvd-package}
\alias{RRedsvd-package}
\alias{RRedsvd}
\docType{package}
\title{
RRedSVD
}
\description{
SVD(singular value decomposition) by using RedSVD
}
\details{
\tabular{ll}{
Package: \tab RRedsvd\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-07-08\cr
License: \tab MPL2\cr
}
~~ An overview of how to use the package, including the most important ~~
"Finding structure with randomness: Stochastic algorithms for
constructing approximate matrix decompositions"
~~ functions ~~
}

\author{
Xiange Dubian
Maintainer: Who to complain to <xiangze gmail.com>
~~ The author and/or maintainer of the package ~~
}
\references{
https://code.google.com/p/redsvd/wiki/English
http://amath.colorado.edu/faculty/martinss/Pubs/2009_HMT_random_review.pdf
http://amath.colorado.edu/faculty/martinss/Talks/2009_NIPS_tutorial.pdf
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
